refactor: move services to their own file
Move services defined in the bootstrap worker to their own file to
better track all of them in one place.
